test(tools): consolidate PrefectWorkerHealthTool tests
